PERFORMANCE - Graph in performance tool is unreadable https://jira.ccpgames.com/browse/EVEDEV-2477
Fixed compile errors on x64

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1184504]

+++ Sprite2d/Tr2Sprite2dLineTrace.cpp
@@ -383,9 +383,9 @@ void Tr2Sprite2dLineTrace::AddSegment(
 		m_drawCalls.back().indexCount = uint16_t( m_renderIndices.size() - m_drawCalls.back().indexOffset );
 
 		DrawCall dc;
-		dc.vertexOffset = m_renderVertices.size();
+		dc.vertexOffset = uint32_t( m_renderVertices.size() );
 		dc.vertexCount = 0;
-		dc.indexOffset = m_renderIndices.size();
+		dc.indexOffset = uint32_t( m_renderIndices.size() );
 		dc.indexCount = 0;
 		m_drawCalls.push_back( dc );
